requirements.txt
boto==2.41.0
boto3==1.3.1
botocore==1.4.34
defusedxml==0.4.1
dj-database-url==0.4.1
Django==1.9.7
django-storages-redux==1.3.2
docutils==0.12
gunicorn==19.6.0
jmespath==0.9.0
oauthlib==1.1.2
Pillow==3.3.0
psycopg2==2.6.1
PyJWT==1.4.0
python-dateutil==2.5.3
python-social-auth==0.2.19
python3-openid==3.0.10
requests==2.10.0
requests-oauthlib==0.6.1
six==1.10.0
-e git+https://github.com/goanpeca/universal-analytics-python.git@4579eaa8b4d4cf5b30ea48a02e5f467c0d254ade#egg=universal_analytics_python
whitenoise==3.2
colour-runner==0.0.4
mock==2.0.0
google-api-python-client==1.5.1
jsonfield
django-compressor==2.0
